Concensus: Testnet permits larger blocks
For testing purposes and performance tuning
it is desirable to be able to test larger amounts
of transactions than would normally make sense
in order to identify bottlenecks for future growth.

Note: This will break concensus, but as it is testnet
it seems fine if during rollout that larger blocks will
fail validation until we get 67%.

Also trivial rename for clarity.