API: change v1.1.0 to v2.0.0
Considering that we are not going to bump API version for every single release of Semux wallet, it makes more sense to let API have its own versioning numbers following semver.
As the v1.1.0 API introduces a few incompatible changes with v1.0.1 API, it should be named as v2.0.0 instead.